% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDis.R
\name{calcDis}
\alias{calcDis}
\title{Calculates d score.}
\usage{
calcDis(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
If input is a numeric vector, a number is returned. For a matrix, a numeric vector is returned.
}
\description{
\emph{d} score of a nucleotide is a measure of dissimilarity of its normalized reactivity scores.
}
\examples{
#Lower standard deviation of reactivites results in lower d-score.
calcDis(rnorm(10, 1, 0.2))
calcDis(rnorm(10, 1, 0.6))
}
