% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.guided.R
\name{dStruct.guided}
\alias{dStruct.guided}
\title{Performs guided discovery of differentially reactive regions.}
\usage{
dStruct.guided(
  rdf,
  reps_A,
  reps_B,
  batches = FALSE,
  within_combs = NULL,
  between_combs = NULL,
  check_quality = TRUE,
  quality = "auto",
  evidence = 0
)
}
\arguments{
\item{rdf}{Dataframe of reactivities for each sample. Each column must be labelled as A1, A2, ..., B1, B2, ...}

\item{reps_A}{Number of replicates of group A.}

\item{reps_B}{Number of replicates of group B.}

\item{batches}{Logical suggesting if replicates of group A and B were performed in batches and are labelled accordingly. If TRUE, a heterogeneous/homogeneous subset may not have multiple samples from the same batch.}

\item{within_combs}{Data.frame with each column containing groupings of replicates of groups A or B, which will be used to assess within-group variation.}

\item{between_combs}{Dataframe with each column containing groupings of replicates of groups A and B, which will be used to assess between-group variation.}

\item{check_quality}{Logical, if TRUE, check regions for quality.}

\item{quality}{Worst allowed quality for a region to be tested.}

\item{evidence}{Minimum evidence of increase in variation from within-group comparisons to between-group comparisons for a region to be tested.}
}
\value{
p-value for the tested region (estimated using one-sided Wilcoxon signed rank test) and the median of nucleotide-wise difference of between-group and within-group d-scores.
}
\description{
Performs guided discovery of differentially reactive regions.
}
\examples{
#Load Wan et al., 2014 data
data(wan2014)

#Run dStruct in the guided mode on first region in wan2014.
dStruct.guided(wan2014[[1]], reps_A = 2, reps_B = 1)
}
