% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContigRegions.R
\name{getContigRegions}
\alias{getContigRegions}
\title{Identifies contiguous regions from a list of nucleotide indices.}
\usage{
getContigRegions(x, gap = 0)
}
\arguments{
\item{x}{A vector of integers.}

\item{gap}{Allowed gap to merge regions.}
}
\value{
Dataframe storing start and stop sites of continguous regions.
}
\description{
Identifies contiguous regions from a list of nucleotide indices.
}
\examples{
#Convert an integer vector of nucleotide positions to a data frame containing the coordinates of contiguous regions.
nucleotide_positions <- c(1, 3, 2, 8, 4:7, 11:20)
getContigRegions(nucleotide_positions)

#Merge regions if their end points are within 3 nt of each other.
getContigRegions(nucleotide_positions, gap = 3)
}
