% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lai2019}
\alias{lai2019}
\title{\emph{Saccharomyces cerevisiae} Structure-seq data}
\format{
An object of class \code{"list"}.
}
\source{
Raw data from \href{https://doi.org/10.1534/genetics.119.302058}{Lai et al., 2019} in processed form.
}
\usage{
data("lai2019")
}
\description{
Data from a Structure-seq assay of five samples of \emph{S. cerevisiae}, three
of which were wild-type samples and two mutant samples. The data was
pre-processed to obtain DMS reactivities as described by Lai et al. (2019).
}
\examples{
data("lai2019")
}
\references{
Lai et al. (2019) Genetics, Vol. 212, 153–174
(\href{https://doi.org/10.1534/genetics.119.302058}{Genetics})
}
\keyword{datasets}
