% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.eight.normalize.R
\name{two.eight.normalize}
\alias{two.eight.normalize}
\title{Normalizes reactivity vector.}
\usage{
two.eight.normalize(raw.estimates)
}
\arguments{
\item{raw.estimates}{A vector of raw reactivities.}
}
\value{
A vector of normalized reactivities.
}
\description{
Normalizes raw reactivities using 2-8 \% method.
}
\examples{
two.eight.normalize(c(NA, rnorm(20, 0.5, 0.3), NA, -999))
}
