% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wan2014}
\alias{wan2014}
\title{\emph{Homo sapiens} PARS data}
\format{
An object of class \code{"list"}.
}
\source{
Counts data from \href{https://doi.org/10.1038/nature12946}{Wan et al., 2014} in processed form.
}
\usage{
data(wan2014)
}
\description{
Data from a PARS assay of a family trio of mother, father, and child. The data was
pre-processed to obtain PARS scores as described in Choudhary et al. (2019).
}
\examples{
data(wan2014)
}
\references{
Wan et al., Nature, 505, 706–709 (2014)
(\href{https://doi.org/10.1038/nature12946}{Nature})
}
\keyword{datasets}
