% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{calcDis}
\alias{calcDis}
\title{Calculates d score.}
\usage{
calcDis(x)
}
\arguments{
\item{x}{A numeric vector or matrix.}
}
\value{
If input is a numeric vector, a number is returned. For a matrix, a numeric vector is returned.
}
\description{
Calculates d score.
}
