% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{dCombs}
\alias{dCombs}
\title{Assesses within-group variation.}
\usage{
dCombs(rdf, combs)
}
\arguments{
\item{rdf}{Dataframe of reactivities for each sample.}

\item{combs}{Data.frame with each column containing groupings of samples.}
}
\value{
Nucleotide-wise d score.
}
\description{
Assesses within-group variation.
}
