% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dCombs.R
\name{dCombs}
\alias{dCombs}
\title{Assesses within-group or between-group variation.}
\usage{
dCombs(rdf, combs)
}
\arguments{
\item{rdf}{Data.frame of reactivities for each sample.}

\item{combs}{Data.frame with each column containing groupings of samples.}
}
\value{
Nucleotide-wise d score.
}
\description{
Assesses within-group or between-group variation.
}
\examples{
#Example of a data frame with reactivities.
reacs <- data.frame(matrix(runif(30, 0, 10), 10, 3))

#The columns of data frame with must indicate sample grouping and id.
colnames(reacs) <- c("A1", "A2", "B1")

#Get nucleotide-wise dissimilarity scores for a set of samples.
dCombs(rdf = reacs, combs = data.frame(c("A1", "B1")))
}
