% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{dStruct.guided}
\alias{dStruct.guided}
\title{Performs guided discovery of differentially reactive regions.}
\usage{
dStruct.guided(rdf, reps_A, reps_B, batches = F, within_combs = NULL,
  between_combs = NULL, check_quality = TRUE, quality = "auto",
  evidence = 0)
}
\arguments{
\item{rdf}{Dataframe of reactivities for each sample.}

\item{reps_A}{Number of replicates of group A.}

\item{reps_B}{Number of replicates of group B.}

\item{batches}{Logical suggesting if replicates of group A and B were performed in batches and are labelled accordingly.}

\item{within_combs}{Data.frame with each column containing groupings of replicates of groups A or B, which will be used to assess within-group variation.}

\item{between_combs}{Dataframe with each column containing groupings of replicates of groups A and B, which will be used to assess between-group variation.}

\item{check_quality}{Logical, if TRUE, check regions for quality.}

\item{quality}{Worst allowed quality for a region to be tested.}

\item{evidence}{Minimum evidence of between-group variation for a region to be tested.}
}
\value{
p-value for the tested region, estimated using one-sided Wilcoxon signed rank test.
}
\description{
Performs guided discovery of differentially reactive regions.
}
