% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{dStruct}
\alias{dStruct}
\title{Performs de novo discovery of differentially reactive regions.}
\usage{
dStruct(rdf, reps_A, reps_B, batches = F, min_length = 11,
  check_signal_strength = T, check_nucs = T, check_quality = T,
  quality = "auto", evidence = 0, signal_strength = 0.1,
  within_combs = NULL, between_combs = NULL, ind_regions = T, gap = 1,
  get_FDR = T, proximity_assisted = F, proximity = 10,
  proximity_defined_length = 30)
}
\arguments{
\item{rdf}{Dataframe of reactivities for each sample.}

\item{reps_A}{Number of replicates of group A.}

\item{reps_B}{Number of replicates of group B.}

\item{batches}{Logical suggesting if replicates of group A and B were performed in batches and are labelled accordingly. If TRUE, a heterogeneous/homogeneous subset may not have multiple samples from the same batch.}

\item{min_length}{Minimum length of constructed regions.}

\item{check_signal_strength}{Logical, if TRUE, construction of regions must be based on nucleotides that have a minimum absolute value of reactivity.}

\item{check_nucs}{Logical, if TRUE, constructed regions must have a minimum number of nucleotides participating in Wilcoxon signed rank test.}

\item{check_quality}{Logical, if TRUE, check constructed regions for quality.}

\item{quality}{Worst allowed quality for a region to be tested.}

\item{evidence}{Minimum evidence of increase in variation from within-group comparisons to between-group comparisons for a region to be tested.}

\item{signal_strength}{Threshold for minimum signal strength.}

\item{within_combs}{Data.frame with each column containing groupings of replicates of groups A or B, which will be used to assess within-group variation.}

\item{between_combs}{Dataframe with each column containing groupings of replicates of groups A and B, which will be used to assess between-group variation.}

\item{ind_regions}{Logical, if TRUE, test each region found in the transcript separately.}

\item{gap}{Integer. Join regions if they are separated by these many nucleotides.}

\item{get_FDR}{Logical, if FALSE, FDR is not reported.}

\item{proximity_assisted}{Logical, if TRUE, proximally located regions are tested together.}

\item{proximity}{Maximum distance between constructed regions for them to be considered proximal.}

\item{proximity_defined_length}{If performing a "proximity-assisted" test, minimum end-to-end length of a region to be tested.}
}
\value{
Constructs regions, reports p-values and FDR for them.
}
\description{
Performs de novo discovery of differentially reactive regions.
}
