% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{dStructome}
\alias{dStructome}
\title{Performs de novo discovery of differentially reactive regions for transcriptome-wide data.}
\usage{
dStructome(rl, reps_A, reps_B, batches = F, min_length = 11,
  check_signal_strength = T, check_nucs = T, check_quality = T,
  quality = "auto", evidence = 0, signal_strength = 0.1,
  within_combs = NULL, between_combs = NULL, ind_regions = T, gap = 1,
  processes = "auto", method = "denovo", proximity_assisted = F,
  proximity = 10)
}
\arguments{
\item{rl}{List of dataframes of reactivities for each sample.}

\item{reps_A}{Number of replicates of group A.}

\item{reps_B}{Number of replicates of group B.}

\item{batches}{Logical suggesting if replicates of group A and B were performed in batches and are labelled accordingly.}

\item{min_length}{Minimum length of constructed regions.}

\item{check_signal_strength}{Logical, if TRUE, constructed regions must have a minimum signal strength.}

\item{check_nucs}{Logical, if TRUE, constructed regions must have a minimum number of nucleotides participating in Wilcoxon signed rank test.}

\item{check_quality}{Logical, if TRUE, check constructed regions for quality.}

\item{quality}{Worst allowed quality for a region to be tested.}

\item{evidence}{Minimum evidence of between-group variation for a region to be tested.}

\item{signal_strength}{Threshold for minimum signal strength.}

\item{within_combs}{Data.frame with each column containing groupings of replicates of groups A or B, which will be used to assess within-group variation.}

\item{between_combs}{Dataframe with each column containing groupings of replicates of groups A and B, which will be used to assess between-group variation.}

\item{ind_regions}{Logical, if TRUE, test each region found in the transcript separately.}

\item{gap}{Integer. Join regions if they are separated by these many nucleotides.}

\item{processes}{Number of parallel processes to use.}

\item{method}{Character specifying either guided or de novo discovery approach.}
}
\value{
Constructs regions, reports p-values and FDR for them.
}
\description{
Performs de novo discovery of differentially reactive regions for transcriptome-wide data.
}
