% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{getContigRegions}
\alias{getContigRegions}
\title{Identifies contiguous regions from a list of nucleotide indices.}
\usage{
getContigRegions(x, gap = 0)
}
\arguments{
\item{x}{A vector of integers.}

\item{gap}{Allowed gap to merge regions.}
}
\value{
Dataframe storing start and stop sites of continguous regions.
}
\description{
Identifies contiguous regions from a list of nucleotide indices.
}
