% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{normalizer}
\alias{normalizer}
\title{Returns normalizer for reactivity vector.}
\usage{
normalizer(raw.estimates)
}
\arguments{
\item{raw.estimates}{A vector of raw reactivities.}
}
\value{
The normalization factor.
}
\description{
Assesses normalization factor for raw reactivities using 2-8 \% method.
}
