% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{plot_dStructurome}
\alias{plot_dStructurome}
\title{Plots differentially reactive regions.}
\usage{
plot_dStructurome(rl, diff_regions, outfile, fdr = 0.05, ylim = c(-0.05, 3))
}
\arguments{
\item{rl}{List of dataframes of reactivities for each sample.}

\item{diff_regions}{Dataframe of regions with significance of differentially reactivity.}

\item{outfile}{The name for pdf file which will be saved.}

\item{fdr}{FDR threshold for plotted regions.}

\item{ylim}{Y-axis limits for plots.}
}
\value{
Saves a PDF for all differentially reactive regions. Returns NULL.
}
\description{
Plots differentially reactive regions.
}
