% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dStruct.R
\name{two.eight.normalize}
\alias{two.eight.normalize}
\title{Normalizes reactivity vector.}
\usage{
two.eight.normalize(raw.estimates)
}
\arguments{
\item{raw.estimates}{A vector of raw reactivities.}
}
\value{
A vector of normalized reactivities.
}
\description{
Normalizes raw reactivities using 2-8 \% method.
}
