% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoEightNormalize.R
\name{twoEightNormalize}
\alias{twoEightNormalize}
\title{Normalizes reactivity vector.}
\usage{
twoEightNormalize(raw.estimates)
}
\arguments{
\item{raw.estimates}{A vector of raw reactivities.}
}
\value{
A vector of normalized reactivities.
}
\description{
Normalizes raw reactivities using 2-8 \% method.
}
\examples{
twoEightNormalize(c(NA, rnorm(20, 0.5, 0.3), NA, -999))
}
